// functions/api/_lib/analytics/types.ts
export type AnalyticsTimeframe =
  | 'last7Days'
  | 'last30Days'
  | 'last90Days'
  | 'yearToDate'
  | 'last12Months'
  | 'lifetime';

export type TimeframeMeta = {
  key: AnalyticsTimeframe;
  label: string;
  compareLabel: string;
  multiplier: number;
  hasComparison?: boolean;
};
// ...
export type AnalyticsDateRange = {
  startDate: string;
  endDate: string;
};

export type AnalyticsTimeframeRange = {
  timeframe: TimeframeMeta;
  current: AnalyticsDateRange;
  previous: AnalyticsDateRange | null;
};